test: merge physics-option input-form tests into one file
test_nested_physics.py (nested family tags, gh#972),
test_physics_families.py (registry + coerce_nested_to_flat) and
test_physics_orthogonal.py (orthogonal form) all exercised the same
physics-option input surface across three sibling modules, with a
duplicated _unwrap helper. They now live in test_physics_options.py
with the bodies unchanged; 143 collected items before and after.
